feat(Views): refactor to View base class and add angularjs support
