+++ packages/cerebral/src/views/angularjs/index.js
@@ -1,0 +1,4 @@
+import module from './module'
+export {default as connect} from './connect'
+
+export default module
