+++ packages/cerebral/src/views/inferno/StateContainer.js
@@ -1,0 +1,16 @@
+import Component from 'inferno-component'
+import {createDummyController} from '../../utils'
+
+class StateContainer extends Component {
+  getChildContext () {
+    const controller = createDummyController(this.props.state, this.props.signals)
+    return {
+      controller: controller
+    }
+  }
+  render () {
+    return this.props.children
+  }
+}
+
+export default StateContainer
