+++ packages/cerebral/src/views/inferno/connect.js
