+++ packages/cerebral/src/views/inferno/index.js
@@ -1,0 +1,7 @@
+import Hoc from './Hoc'
+import connectFactory, { decoratorFactory } from './connect'
+
+export {default as Container} from './Container'
+export {default as StateContainer} from './StateContainer'
+export const connect = connectFactory(Hoc)
+export const decorator = decoratorFactory(Hoc)
