+++ packages/cerebral/src/views/react/react.test.js
@@ -13,12 +13,12 @@ const React = require('react')
 const ReactDOM = require('react-dom')
 const TestUtils = require('react-addons-test-utils')
 const assert = require('assert')
-const Controller = require('../Controller').default
-const {state, signal, props} = require('../tags')
-const {compute} = require('../')
-const {Container, StateContainer, connect, decorator} = require('./react')
+const Controller = require('../../Controller').default
+const {state, signal, props} = require('../../tags')
+const {compute} = require('../../')
+const {Container, StateContainer, connect, decorator} = require('./')
 
-describe('React', () => {
+describe.only('React', () => {
   describe('state container', () => {
     it('should be able to wrap app with container', () => {
       class TestComponent extends React.Component {
@@ -332,7 +332,7 @@ describe('React', () => {
           <TestComponent />
         </Container>
       ))
-      assert.equal(TestUtils.findRenderedComponentWithType(tree, TestComponent)._hasWarnedBigComponent, true)
+      assert.equal(TestUtils.findRenderedComponentWithType(tree, TestComponent).view._hasWarnedBigComponent, true)
       assert.equal(warnCount, 1)
     })
     it('should throw an error if no container component provided', () => {
